verify email and create user in one transaction
