+++ Makefile
@@ -8,7 +8,7 @@ createdb:
 sleep10:
 	sleep 10
 
-local: postgres sleep10 createdb server
+local: redis postgres sleep10 createdb server
 
 migrateup:
 	migrate -path db/migrations -database "postgresql://root:secret@localhost:5432/simple_bank?sslmode=disable" -verbose up
@@ -57,6 +57,7 @@ evans:
 	evans --host localhost --port 9090 -r repl
 
 redis:
+	docker rm -f redis
 	docker run --name redis -p 6379:6379 -d redis:7.4-alpine
 
 .PHONY: createdb postgres createdb migrateup migratedown migrateup1 migratedown1 dropdb sqlc test server mock db_docs db_schema proto evans local sleep10 alpine
