+++ db/sqlc/store.go
@@ -9,6 +9,7 @@ import (
 type Store interface {
 	Querier
 	TransferTx(ctx context.Context, arg TransferTxParams) (TransferTxResult, error)
+	CreateUserTx(ctx context.Context, arg CreateUserTxParams) (CreateUserTxResult, error)
 }
 
 type SQLStore struct {
@@ -39,84 +40,4 @@ func (store *SQLStore) execTx(ctx context.Context, fn func(queries *Queries) err
 		return err
 	}
 	return tx.Commit()
-}
-
-type TransferTxParams struct {
-	FromAccountId int64 `json:"from_account_id"`
-	ToAccountId   int64 `json:"to_account_id"`
-	Amount        int64 `json:"amount"`
-}
-
-type TransferTxResult struct {
-	Transfer    Transfer `json:"transfer"`
-	FromAccount Account  `json:"from_account"`
-	ToAccount   Account  `json:"to_account"`
-	FromEntry   Entry    `json:"from_entry"`
-	ToEntry     Entry    `json:"to_entry"`
-}
-
-func (store *SQLStore) TransferTx(ctx context.Context, arg TransferTxParams) (TransferTxResult, error) {
-	var result TransferTxResult
-
-	err := store.execTx(ctx, func(queries *Queries) error {
-		var err error
-
-		//txName := ctx.Value(txKey)
-
-		//fmt.Println(txName, "Create Transfer")
-		result.Transfer, err = queries.CreateTransfer(ctx, CreateTransferParams{
-			FromAccountID: arg.FromAccountId,
-			ToAccountID:   arg.ToAccountId,
-			Amount:        arg.Amount,
-		})
-		if err != nil {
-			return err
-		}
-
-		//fmt.Println(txName, "Create Entry 1")
-		result.FromEntry, err = queries.CreateEntry(ctx, CreateEntryParams{
-			AccountID: arg.FromAccountId,
-			Amount:    -arg.Amount,
-		})
-		if err != nil {
-			return err
-		}
-
-		//fmt.Println(txName, "Create Entry 2")
-		result.ToEntry, err = queries.CreateEntry(ctx, CreateEntryParams{
-			AccountID: arg.ToAccountId,
-			Amount:    arg.Amount,
-		})
-		if err != nil {
-			return err
-		}
-
-		//fmt.Println(txName, "update account 1")
-		if arg.FromAccountId < arg.ToAccountId {
-			result.FromAccount, result.ToAccount, err = addMoney(ctx, queries, arg.FromAccountId, -arg.Amount, arg.ToAccountId, arg.Amount)
-		} else {
-			result.ToAccount, result.FromAccount, err = addMoney(ctx, queries, arg.ToAccountId, arg.Amount, arg.FromAccountId, -arg.Amount)
-		}
-
-		return err
-	})
-	return result, err
-}
-
-func addMoney(ctx context.Context, q *Queries, accountID1, amount1, accountId2, amount2 int64) (account1, account2 Account, err error) {
-	account1, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
-		Amount: amount1,
-		ID:     accountID1,
-	})
-	if err != nil {
-		return
-	}
-
-	account2, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
-		Amount: amount2,
-		ID:     accountId2,
-	})
-
-	return
-
 }
