+++ gapi/rpc_create_user.go
@@ -24,14 +24,32 @@ func (s *Server) CreateUser(ctx context.Context, req *pb.CreateUserRequest) (*pb
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "failed to hash password: %v", err)
 	}
-	arg := db.CreateUserParams{
-		Username:       req.GetUsername(),
-		HashedPassword: hashedPassword,
-		FullName:       req.GetFullName(),
-		Email:          req.GetEmail(),
+	arg := db.CreateUserTxParams{
+		CreateUserParams: db.CreateUserParams{
+			Username:       req.GetUsername(),
+			HashedPassword: hashedPassword,
+			FullName:       req.GetFullName(),
+			Email:          req.GetEmail(),
+		},
+		AfterCreate: func(user db.User) error {
+			opts := []asynq.Option{
+				asynq.MaxRetry(10),
+				asynq.ProcessIn(10 * time.Second),
+				asynq.Queue(worker.QueueCritical),
+			}
+
+			err = s.taskDistributor.DistributeTaskSendVerifyEmail(
+				ctx,
+				&worker.PayloadSendVerifyEmail{
+					Username: user.Username,
+				},
+				opts...,
+			)
+			return err
+		},
 	}
 
-	user, err := s.store.CreateUser(ctx, arg)
+	txResult, err := s.store.CreateUserTx(ctx, arg)
 	if err != nil {
 		if pgError, ok := err.(*pq.Error); ok {
 			switch pgError.Code.Name() {
@@ -42,26 +60,8 @@ func (s *Server) CreateUser(ctx context.Context, req *pb.CreateUserRequest) (*pb
 		return nil, status.Errorf(codes.Internal, "failed to create user: %s", err)
 	}
 
-	opts := []asynq.Option{
-		asynq.MaxRetry(10),
-		asynq.ProcessIn(10 * time.Second),
-		asynq.Queue(worker.QueueCritical),
-	}
-
-	err = s.taskDistributor.DistributeTaskSendVerifyEmail(
-		ctx,
-		&worker.PayloadSendVerifyEmail{
-			Username: user.Username,
-		},
-		opts...,
-	)
-
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "failed to distribute task to send  verify email: %v", err)
-	}
-
 	res := &pb.CreateUserResponse{
-		User: convertUser(user),
+		User: convertUser(txResult.User),
 	}
 	return res, nil
 }
